// src/Types/interface.ts
export interface Species {
    name: string;
    url:  string;
}
// ...
export interface Stat {
    base_stat: number;
    effort:    number;
    stat:      Species;
}